rel as nodes fixture
